+++ tests/BaseDriverTest.php
@@ -105,6 +105,19 @@ abstract class BaseDriverTest extends PHPUnit_Framework_TestCase
         $this->redisProxy->info('unknown');
     }
 
+    public function testDbSize()
+    {
+        self::assertEquals(0, $this->redisProxy->dbsize());
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertEquals(1, $this->redisProxy->dbsize());
+        for ($i = 0; $i < 9; ++$i) {
+            self::assertTrue($this->redisProxy->set("my_key_$i", "my_value$i"));
+        }
+        self::assertEquals(10, $this->redisProxy->dbsize());
+        $keys = $this->redisProxy->keys('*');
+        self::assertEquals(count($keys), $this->redisProxy->dbsize());
+    }
+
     public function testKeys()
     {
         $keys = $this->redisProxy->keys('*');
@@ -126,26 +139,153 @@ abstract class BaseDriverTest extends PHPUnit_Framework_TestCase
         $this->redisProxy->keys();
     }
 
-    public function testDbSize()
-    {
-        self::assertEquals(0, $this->redisProxy->dbsize());
-        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
-        self::assertEquals(1, $this->redisProxy->dbsize());
-        for ($i = 0; $i < 9; ++$i) {
-            self::assertTrue($this->redisProxy->set("my_key_$i", "my_value$i"));
-        }
-        self::assertEquals(10, $this->redisProxy->dbsize());
-        $keys = $this->redisProxy->keys('*');
-        self::assertEquals(count($keys), $this->redisProxy->dbsize());
-    }
-
-    public function testSetGet()
+    public function testSetAndGet()
     {
         self::assertNull($this->redisProxy->get('my_key'));
         self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
         self::assertEquals('my_value', $this->redisProxy->get('my_key'));
         self::assertTrue($this->redisProxy->set('my_key', 'my_new_value'));
         self::assertEquals('my_new_value', $this->redisProxy->get('my_key'));
+    }
+
+    public function testExists()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->exists('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertEquals('my_value', $this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->exists('my_key'));
+    }
+
+    public function testDumpAndRestore()
+    {
+        self::assertNull($this->redisProxy->get('my_key_1'));
+        self::assertNull($this->redisProxy->get('my_key_2'));
+        self::assertNull($this->redisProxy->dump('my_key_1'));
+        self::assertNull($this->redisProxy->dump('my_key_2'));
+        self::assertTrue($this->redisProxy->set('my_key_1', 'my_value_1'));
+        self::assertTrue($this->redisProxy->set('my_key_2', 'my_value_2'));
+        self::assertEquals('my_value_1', $this->redisProxy->get('my_key_1'));
+        self::assertEquals('my_value_2', $this->redisProxy->get('my_key_2'));
+        $value1 = $this->redisProxy->dump('my_key_1');
+        $value2 = $this->redisProxy->dump('my_key_2');
+        self::assertEquals(2, $this->redisProxy->del('my_key_1', 'my_key_2'));
+        self::assertNull($this->redisProxy->get('my_key_1'));
+        self::assertNull($this->redisProxy->get('my_key_2'));
+        self::assertTrue($this->redisProxy->restore('my_key_1', 0, $value2));
+        self::assertTrue($this->redisProxy->restore('my_key_2', 0, $value1));
+        self::assertEquals('my_value_2', $this->redisProxy->get('my_key_1'));
+        self::assertEquals('my_value_1', $this->redisProxy->get('my_key_2'));
+    }
+
+    public function testExpire()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->expire('my_key', 10));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertTrue($this->redisProxy->expire('my_key', 10));
+    }
+
+    public function testPexpire()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->pexpire('my_key', 10000));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertTrue($this->redisProxy->pexpire('my_key', 10000));
+    }
+
+    public function testTtl()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertEquals(-2, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->expire('my_key', 10));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+
+        self::assertNull($this->redisProxy->get('my_key_2'));
+        self::assertEquals(-2, $this->redisProxy->ttl('my_key_2'));
+        self::assertTrue($this->redisProxy->set('my_key_2', 'my_value'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key_2'));
+        self::assertTrue($this->redisProxy->pexpire('my_key_2', 10000));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key_2'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key_2'));
+    }
+
+    public function testExpireat()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->expireat('my_key', time() + 10));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertTrue($this->redisProxy->expireat('my_key', time() + 10));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+    }
+
+    public function testPexpireat()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->pexpireat('my_key', (time() + 10) * 1000));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertTrue($this->redisProxy->pexpireat('my_key', (time() + 10) * 1000));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->pttl('my_key'));
+        self::assertLessThanOrEqual(10000, $this->redisProxy->pttl('my_key'));
+    }
+
+    public function testPersist()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->persist('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertFalse($this->redisProxy->persist('my_key'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->expire('my_key', 10));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->persist('my_key'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key'));
+    }
+
+    public function testGetSet()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertNull($this->redisProxy->getset('my_key', 'my_value'));
+        self::assertEquals('my_value', $this->redisProxy->getset('my_key', 'my_new_value'));
+        self::assertEquals('my_new_value', $this->redisProxy->get('my_key'));
+    }
+
+    public function testSetnx()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->setnx('my_key', 'my_value'));
+        self::assertEquals('my_value', $this->redisProxy->get('my_key'));
+        self::assertFalse($this->redisProxy->setnx('my_key', 'my_new_value'));
+        self::assertEquals('my_value', $this->redisProxy->get('my_key'));
+    }
+
+    public function testSetex()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->setex('my_key', 10, 'my_value'));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+    }
+
+    public function testPsetex()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 'my_value'));
+        self::assertEquals(-1, $this->redisProxy->ttl('my_key'));
+        self::assertTrue($this->redisProxy->psetex('my_key', 10000, 'my_value'));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->ttl('my_key'));
+        self::assertLessThanOrEqual(10, $this->redisProxy->ttl('my_key'));
+        self::assertGreaterThanOrEqual(0, $this->redisProxy->pttl('my_key'));
+        self::assertLessThanOrEqual(10000, $this->redisProxy->pttl('my_key'));
     }
 
     public function testMset()
@@ -288,6 +428,94 @@ abstract class BaseDriverTest extends PHPUnit_Framework_TestCase
     public function testDeleteNoKeys()
     {
         $this->redisProxy->delete();
+    }
+
+    public function testIncr()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertEquals(1, $this->redisProxy->incr('my_key'));
+        self::assertEquals(1, $this->redisProxy->get('my_key'));
+        self::assertEquals(2, $this->redisProxy->incr('my_key', 5));
+        self::assertEquals(2, $this->redisProxy->get('my_key'));
+        self::assertEquals(3, $this->redisProxy->incr('my_key', -4.1234));
+        self::assertEquals(3, $this->redisProxy->get('my_key'));
+    }
+
+    public function testIncrby()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertEquals(1, $this->redisProxy->incrby('my_key'));
+        self::assertEquals(1, $this->redisProxy->get('my_key'));
+        self::assertEquals(6, $this->redisProxy->incrby('my_key', 5));
+        self::assertEquals(6, $this->redisProxy->get('my_key'));
+        self::assertEquals(10, $this->redisProxy->incrby('my_key', 4.1234));
+        self::assertEquals(10, $this->redisProxy->get('my_key'));
+        self::assertEquals(5, $this->redisProxy->incrby('my_key', -5));
+        self::assertEquals(5, $this->redisProxy->get('my_key'));
+        self::assertEquals(0, $this->redisProxy->incrby('my_key', -5));
+        self::assertEquals(0, $this->redisProxy->get('my_key'));
+    }
+
+    public function testIncrbyfloat()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertEquals(1, $this->redisProxy->incrbyfloat('my_key'));
+        self::assertEquals(1, $this->redisProxy->get('my_key'));
+        self::assertEquals(6, $this->redisProxy->incrbyfloat('my_key', 5));
+        self::assertEquals(6, $this->redisProxy->get('my_key'));
+        self::assertEquals(10.1234, $this->redisProxy->incrbyfloat('my_key', 4.1234));
+        self::assertEquals(10.1234, $this->redisProxy->get('my_key'));
+        self::assertEquals(5.1234, $this->redisProxy->incrbyfloat('my_key', -5));
+        self::assertEquals(5.1234, $this->redisProxy->get('my_key'));
+        self::assertEquals(-1.4, $this->redisProxy->incrbyfloat('my_key', -6.5234));
+        self::assertEquals(-1.4, $this->redisProxy->get('my_key'));
+    }
+
+    public function testDecr()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 10));
+        self::assertEquals(10, $this->redisProxy->get('my_key'));
+        self::assertEquals(9, $this->redisProxy->decr('my_key'));
+        self::assertEquals(9, $this->redisProxy->get('my_key'));
+        self::assertEquals(8, $this->redisProxy->decr('my_key', 5));
+        self::assertEquals(8, $this->redisProxy->get('my_key'));
+        self::assertEquals(7, $this->redisProxy->decr('my_key', -4.1234));
+        self::assertEquals(7, $this->redisProxy->get('my_key'));
+    }
+
+    public function testDecrby()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 10));
+        self::assertEquals(10, $this->redisProxy->get('my_key'));
+        self::assertEquals(9, $this->redisProxy->decrby('my_key'));
+        self::assertEquals(9, $this->redisProxy->get('my_key'));
+        self::assertEquals(4, $this->redisProxy->decrby('my_key', 5));
+        self::assertEquals(4, $this->redisProxy->get('my_key'));
+        self::assertEquals(0, $this->redisProxy->decrby('my_key', 4.1234));
+        self::assertEquals(0, $this->redisProxy->get('my_key'));
+        self::assertEquals(5, $this->redisProxy->decrby('my_key', -5));
+        self::assertEquals(5, $this->redisProxy->get('my_key'));
+        self::assertEquals(10, $this->redisProxy->decrby('my_key', -5));
+        self::assertEquals(10, $this->redisProxy->get('my_key'));
+    }
+
+    public function testDecrbyfloat()
+    {
+        self::assertNull($this->redisProxy->get('my_key'));
+        self::assertTrue($this->redisProxy->set('my_key', 10));
+        self::assertEquals(10, $this->redisProxy->get('my_key'));
+        self::assertEquals(9, $this->redisProxy->decrbyfloat('my_key'));
+        self::assertEquals(9, $this->redisProxy->get('my_key'));
+        self::assertEquals(4, $this->redisProxy->decrbyfloat('my_key', 5));
+        self::assertEquals(4, $this->redisProxy->get('my_key'));
+        self::assertEquals(-0.1234, $this->redisProxy->decrbyfloat('my_key', 4.1234));
+        self::assertEquals(-0.1234, $this->redisProxy->get('my_key'));
+        self::assertEquals(4.8766, $this->redisProxy->decrbyfloat('my_key', -5));
+        self::assertEquals(4.8766, $this->redisProxy->get('my_key'));
+        self::assertEquals(11.4, $this->redisProxy->decrbyfloat('my_key', -6.5234));
+        self::assertEquals(11.4, $this->redisProxy->get('my_key'));
     }
 
     public function testScan()
@@ -689,5 +917,254 @@ abstract class BaseDriverTest extends PHPUnit_Framework_TestCase
         }
         self::assertEquals(271, $count);
         self::assertEquals(0, $iterator);
+    }
+
+    public function testLlen()
+    {
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals(1, $this->redisProxy->lpush('my_list_key', 'my_value'));
+        self::assertEquals(1, $this->redisProxy->llen('my_list_key'));
+    }
+
+    public function testLpush()
+    {
+        // add one element
+        self::assertEquals(1, $this->redisProxy->lpush('my_list_key', 'element'));
+        self::assertEquals(1, $this->redisProxy->llen('my_list_key'));
+
+        // add three elements
+        self::assertEquals(4, $this->redisProxy->lpush('my_list_key', 'element_2', 'element_3', 'element_4'));
+        self::assertEquals(4, $this->redisProxy->llen('my_list_key'));
+
+        // add elements which are already in list
+        self::assertEquals(7, $this->redisProxy->lpush('my_list_key', 'element_2', 'element_3', 'element_4'));
+        self::assertEquals(7, $this->redisProxy->llen('my_list_key'));
+
+        // add some new elements and some elements which are already in list
+        self::assertEquals(11, $this->redisProxy->lpush('my_list_key', 'element_2', 'element_3', 'element_5', 'element_6'));
+        self::assertEquals(11, $this->redisProxy->llen('my_list_key'));
+
+        // add three elements as array
+        self::assertEquals(14, $this->redisProxy->lpush('my_list_key', ['element_7', 'element_8', 'element_9']));
+        self::assertEquals(14, $this->redisProxy->llen('my_list_key'));
+    }
+
+    public function testRpush()
+    {
+        // add one element
+        self::assertEquals(1, $this->redisProxy->rpush('my_list_key', 'element'));
+        self::assertEquals(1, $this->redisProxy->llen('my_list_key'));
+
+        // add three elements
+        self::assertEquals(4, $this->redisProxy->rpush('my_list_key', 'element_2', 'element_3', 'element_4'));
+        self::assertEquals(4, $this->redisProxy->llen('my_list_key'));
+
+        // add elements which are already in list
+        self::assertEquals(7, $this->redisProxy->rpush('my_list_key', 'element_2', 'element_3', 'element_4'));
+        self::assertEquals(7, $this->redisProxy->llen('my_list_key'));
+
+        // add some new elements and some elements which are already in list
+        self::assertEquals(11, $this->redisProxy->rpush('my_list_key', 'element_2', 'element_3', 'element_5', 'element_6'));
+        self::assertEquals(11, $this->redisProxy->llen('my_list_key'));
+
+        // add three elements as array
+        self::assertEquals(14, $this->redisProxy->rpush('my_list_key', ['element_7', 'element_8', 'element_9']));
+        self::assertEquals(14, $this->redisProxy->llen('my_list_key'));
+    }
+
+    public function testLrange()
+    {
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertCount(0, $this->redisProxy->lrange('my_list_key', 0, -1));
+        self::assertCount(0, $this->redisProxy->lrange('my_list_key', 0, 2));
+
+        self::assertEquals(2, $this->redisProxy->lpush('my_list_key', 'element_1', 'element_2'));
+        self::assertEquals(2, $this->redisProxy->llen('my_list_key'));
+        self::assertCount(2, $this->redisProxy->lrange('my_list_key', 0, -1));
+        self::assertCount(2, $this->redisProxy->lrange('my_list_key', 0, 2));
+        self::assertEquals(['element_2', 'element_1'], $this->redisProxy->lrange('my_list_key', 0, -1));
+
+        self::assertEquals(4, $this->redisProxy->rpush('my_list_key', 'element_3', 'element_4'));
+        self::assertEquals(4, $this->redisProxy->llen('my_list_key'));
+        self::assertCount(4, $this->redisProxy->lrange('my_list_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->lrange('my_list_key', 0, 2));
+        self::assertEquals(['element_2', 'element_1', 'element_3', 'element_4'], $this->redisProxy->lrange('my_list_key', 0, -1));
+    }
+
+    public function testLindex()
+    {
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 0));
+
+        self::assertEquals(2, $this->redisProxy->lpush('my_list_key', 'element_1', 'element_2'));
+        self::assertEquals(2, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_2', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertEquals('element_1', $this->redisProxy->lindex('my_list_key', 1));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 2));
+
+        self::assertEquals(4, $this->redisProxy->rpush('my_list_key', 'element_3', 'element_4'));
+        self::assertEquals(4, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_2', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertEquals('element_1', $this->redisProxy->lindex('my_list_key', 1));
+        self::assertEquals('element_3', $this->redisProxy->lindex('my_list_key', 2));
+        self::assertEquals('element_4', $this->redisProxy->lindex('my_list_key', 3));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 4));
+    }
+
+    public function testLpop()
+    {
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 0));
+        self::assertNull($this->redisProxy->lpop('my_list_key'));
+        self::assertEquals(2, $this->redisProxy->lpush('my_list_key', 'element_1', 'element_2'));
+        self::assertEquals(2, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_2', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertEquals('element_1', $this->redisProxy->lindex('my_list_key', 1));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 2));
+        self::assertEquals('element_2', $this->redisProxy->lpop('my_list_key'));
+        self::assertEquals(1, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_1', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 1));
+        self::assertEquals('element_1', $this->redisProxy->lpop('my_list_key'));
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 0));
+    }
+
+    public function testRpop()
+    {
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 0));
+        self::assertNull($this->redisProxy->rpop('my_list_key'));
+        self::assertEquals(2, $this->redisProxy->lpush('my_list_key', 'element_1', 'element_2'));
+        self::assertEquals(2, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_2', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertEquals('element_1', $this->redisProxy->lindex('my_list_key', 1));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 2));
+        self::assertEquals('element_1', $this->redisProxy->rpop('my_list_key'));
+        self::assertEquals(1, $this->redisProxy->llen('my_list_key'));
+        self::assertEquals('element_2', $this->redisProxy->lindex('my_list_key', 0));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 1));
+        self::assertEquals('element_2', $this->redisProxy->rpop('my_list_key'));
+        self::assertEquals(0, $this->redisProxy->llen('my_list_key'));
+        self::assertNull($this->redisProxy->lindex('my_list_key', 0));
+    }
+
+    public function testZcard()
+    {
+        self::assertEquals(0, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertEquals(1, $this->redisProxy->zadd('my_sorted_set_key', 1, 'my_member'));
+        self::assertEquals(1, $this->redisProxy->zcard('my_sorted_set_key'));
+    }
+
+    public function testZadd()
+    {
+        self::assertEquals(0, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertEquals(1, $this->redisProxy->zadd('my_sorted_set_key', 1, 'my_member'));
+        self::assertEquals(1, $this->redisProxy->zcard('my_sorted_set_key'));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', 2, 'my_member_2', 3, 'my_member_3'));
+        self::assertEquals(3, $this->redisProxy->zcard('my_sorted_set_key'));
+
+        self::assertEquals(0, $this->redisProxy->zadd('my_sorted_set_key', 10, 'my_member'));
+        self::assertEquals(3, $this->redisProxy->zcard('my_sorted_set_key'));
+
+        self::assertEquals(1, $this->redisProxy->zadd('my_sorted_set_key', 2, 'my_member_2', 3, 'my_member_3', 4, 'my_member_4'));
+        self::assertEquals(4, $this->redisProxy->zcard('my_sorted_set_key'));
+
+        self::assertEquals(3, $this->redisProxy->zadd('my_sorted_set_key', ['my_member_5' => 5, 'my_member_6' => 4, 'my_member_7' => 3]));
+        self::assertEquals(7, $this->redisProxy->zcard('my_sorted_set_key'));
+    }
+
+    public function testZrange()
+    {
+        self::assertEquals(0, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(0, $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertCount(0, $this->redisProxy->zrange('my_sorted_set_key', 0, 2));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', 1, 'element_1', 2, 'element_2'));
+        self::assertEquals(2, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(2, $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertCount(2, $this->redisProxy->zrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_1', 'element_2'], $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_1' => 1, 'element_2' => 2], $this->redisProxy->zrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', 3, 'element_3', 4, 'element_4'));
+        self::assertEquals(4, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(4, $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_1', 'element_2', 'element_3', 'element_4'], $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_1' => 1, 'element_2' => 2, 'element_3' => 3, 'element_4' => 4], $this->redisProxy->zrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', ['element_5' => -5, 'element_6' => -6]));
+        self::assertEquals(6, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(6, $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_6', 'element_5', 'element_1', 'element_2', 'element_3', 'element_4'], $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_6' => -6, 'element_5' => -5, 'element_1' => 1, 'element_2' => 2, 'element_3' => 3, 'element_4' => 4], $this->redisProxy->zrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', ['element_7' => -5, 'element_8' => -6]));
+        self::assertEquals(8, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(8, $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_6', 'element_8', 'element_5', 'element_7', 'element_1', 'element_2', 'element_3', 'element_4'], $this->redisProxy->zrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_6' => -6, 'element_8' => -6, 'element_5' => -5, 'element_7' => -5, 'element_1' => 1, 'element_2' => 2, 'element_3' => 3, 'element_4' => 4], $this->redisProxy->zrange('my_sorted_set_key', 0, -1, true));
+    }
+
+    public function testZrevrange()
+    {
+        self::assertEquals(0, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(0, $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertCount(0, $this->redisProxy->zrevrange('my_sorted_set_key', 0, 2));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', 1, 'element_1', 2, 'element_2'));
+        self::assertEquals(2, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(2, $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertCount(2, $this->redisProxy->zrevrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_2', 'element_1'], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_2' => 2, 'element_1' => 1], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', 3, 'element_3', 4, 'element_4'));
+        self::assertEquals(4, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(4, $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrevrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_4', 'element_3', 'element_2', 'element_1'], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_4' => 4, 'element_3' => 3, 'element_2' => 2, 'element_1' => 1], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', ['element_5' => -5, 'element_6' => -6]));
+        self::assertEquals(6, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(6, $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrevrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_4', 'element_3', 'element_2', 'element_1', 'element_5', 'element_6'], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_4' => 4, 'element_3' => 3, 'element_2' => 2, 'element_1' => 1, 'element_5' => -5, 'element_6' => -6], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1, true));
+
+        self::assertEquals(2, $this->redisProxy->zadd('my_sorted_set_key', ['element_7' => -5, 'element_8' => -6]));
+        self::assertEquals(8, $this->redisProxy->zcard('my_sorted_set_key'));
+        self::assertCount(8, $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertCount(3, $this->redisProxy->zrevrange('my_sorted_set_key', 0, 2));
+        self::assertEquals(['element_4', 'element_3', 'element_2', 'element_1', 'element_7', 'element_5', 'element_8', 'element_6'], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1));
+        self::assertEquals(['element_4' => 4, 'element_3' => 3, 'element_2' => 2, 'element_1' => 1, 'element_7' => -5, 'element_5' => -5, 'element_8' => -6, 'element_6' => -6], $this->redisProxy->zrevrange('my_sorted_set_key', 0, -1, true));
+    }
+
+    public function testType()
+    {
+        self::assertNull($this->redisProxy->type('my_key'));
+        $this->redisProxy->set('my_key', 'my_value');
+        self::assertEquals(RedisProxy::TYPE_STRING, $this->redisProxy->type('my_key'));
+
+        self::assertNull($this->redisProxy->type('my_set_key'));
+        $this->redisProxy->sadd('my_set_key', 'my_member');
+        self::assertEquals(RedisProxy::TYPE_SET, $this->redisProxy->type('my_set_key'));
+
+        self::assertNull($this->redisProxy->type('my_hash_key'));
+        $this->redisProxy->hset('my_hash_key', 'my_key', 'my_value');
+        self::assertEquals(RedisProxy::TYPE_HASH, $this->redisProxy->type('my_hash_key'));
+
+        self::assertNull($this->redisProxy->type('my_list_key'));
+        $this->redisProxy->lpush('my_list_key', 'my_value');
+        self::assertEquals(RedisProxy::TYPE_LIST, $this->redisProxy->type('my_list_key'));
+
+        self::assertNull($this->redisProxy->type('my_sorted_set_key'));
+        $this->redisProxy->zadd('my_sorted_set_key', 1, 'my_value');
+        self::assertEquals(RedisProxy::TYPE_SORTED_SET, $this->redisProxy->type('my_sorted_set_key'));
     }
 }